:boom: change: from js to ts (client)
