+++ client/src/components/Asahi.tsx
@@ -1,35 +1,38 @@
-import * as React from 'react';
+import React from 'react';
 import { useState, useEffect } from 'react';
 import { Loader } from './Loader';
 import { Card } from './Card';
 import '../App.css';
-import yahooNews from './img/yahooNews.png';
+import asahi from './img/asahi.png'; //modify here
 
-export const Yahoo = () => {
-  const [data, setData] = useState([]);
+//modify component Name
+export const Asahi = () => {
+  const [data, setData] = useState<{ [key: string]: string }[]>([]);
 
+  //modify path
   const getData = () => {
-    return fetch('./api/yahoo').then((response) => response.json());
+    return fetch('./api/asahi').then((response) => response.json());
   };
 
   useEffect(() => {
     getData().then((response) => setData(response));
   }, []);
 
+  //modify section text
   if (!data?.length) {
     return <Loader />;
   }
   return (
     <>
       <div className="inner">
-        <h2 className="section">Yahoo! ニュース</h2>
+        <h2 className="section">朝日新聞</h2>
         <div className="container">
           {data?.map((data) => (
             <Card
               key={data.textContent}
               title={data.textContent}
               link={data.href}
-              img={yahooNews}
+              img={data.src || asahi}
               target="_blank"
               fontSize="1.5rem"
               fontWeight="normal"
