+++ client/src/components/Card.tsx
@@ -2,8 +2,12 @@ import React from 'react';
 import './css/card.css';
 import '../reset.css';
 import news from './img/news.png';
+import { TODO_FROM_JS_TO_TS } from '../types';
+interface CardProps {
+  [key: string]: TODO_FROM_JS_TO_TS;
+}
 
-export const Card = (props) => {
+export const Card = (props: CardProps) => {
   const { title, link, img, target, fontSize, fontWeight } = props;
   const style = { fontSize, fontWeight };
   return (
