+++ client/src/components/Excite.tsx
@@ -1,38 +1,36 @@
-import * as React from 'react';
+import React from 'react';
 import { useState, useEffect } from 'react';
 import { Loader } from './Loader';
 import { Card } from './Card';
 import '../App.css';
-import sankei from './img/sankei.png'; //modify here
+import exciteNews from './img/exciteNews.png'; //modify here
 
-//modify component Name
-export const Sankei = () => {
-  const [data, setData] = useState([]);
+export const Excite = () => {
+  //modify here
+  const [data, setData] = useState<{ [key: string]: string }[]>([]);
 
-  //modify path
   const getData = () => {
-    return fetch('./api/sankei').then((response) => response.json());
+    return fetch('./api/excite').then((response) => response.json());
   };
 
   useEffect(() => {
     getData().then((response) => setData(response));
   }, []);
 
-  //modify section text
   if (!data?.length) {
     return <Loader />;
   }
   return (
     <>
       <div className="inner">
-        <h2 className="section">産経新聞</h2>
+        <h2 className="section">excite ニュース</h2>
         <div className="container">
           {data?.map((data) => (
             <Card
               key={data.textContent}
               title={data.textContent}
               link={data.href}
-              img={data.src}
+              img={exciteNews}
               target="_blank"
               fontSize="1.5rem"
               fontWeight="normal"
