+++ client/src/components/Livedoor.tsx
@@ -1,14 +1,15 @@
-import * as React from 'react';
+import React from 'react';
 import { useState, useEffect } from 'react';
 import { Loader } from './Loader';
 import { Card } from './Card';
 import '../App.css';
+import livedoorNews from './img/livedoorNews.png';
 
-export const Nhk = () => {
-  const [data, setData] = useState([]);
+export const Livedoor = () => {
+  const [data, setData] = useState<{ [key: string]: string }[]>([]);
 
   const getData = () => {
-    return fetch('./api/nhk').then((response) => response.json());
+    return fetch('./api/livedoor').then((response) => response.json());
   };
 
   useEffect(() => {
@@ -21,14 +22,14 @@ export const Nhk = () => {
   return (
     <>
       <div className="inner">
-        <h2 className="section">NHK ニュース</h2>
+        <h2 className="section">Livedoor ニュース</h2>
         <div className="container">
           {data?.map((data) => (
             <Card
               key={data.textContent}
               title={data.textContent}
               link={data.href}
-              img={data.src}
+              img={livedoorNews}
               target="_blank"
               fontSize="1.5rem"
               fontWeight="normal"
