+++ client/src/components/Mainichi.tsx
@@ -1,36 +1,38 @@
-import * as React from 'react';
+import React from 'react';
 import { useState, useEffect } from 'react';
 import { Loader } from './Loader';
 import { Card } from './Card';
 import '../App.css';
-import exciteNews from './img/exciteNews.png'; //modify here
+import mainichi from './img/mainichi.png'; //modify here
 
-export const Excite = () => {
-  //modify here
-  const [data, setData] = useState([]);
+//modify component Name
+export const Mainichi = () => {
+  const [data, setData] = useState<{ [key: string]: string }[]>([]);
 
+  //modify path
   const getData = () => {
-    return fetch('./api/excite').then((response) => response.json());
+    return fetch('./api/mainichi').then((response) => response.json());
   };
 
   useEffect(() => {
     getData().then((response) => setData(response));
   }, []);
 
+  //modify section text
   if (!data?.length) {
     return <Loader />;
   }
   return (
     <>
       <div className="inner">
-        <h2 className="section">excite ニュース</h2>
+        <h2 className="section">毎日新聞</h2>
         <div className="container">
           {data?.map((data) => (
             <Card
               key={data.textContent}
               title={data.textContent}
               link={data.href}
-              img={exciteNews}
+              img={mainichi}
               target="_blank"
               fontSize="1.5rem"
               fontWeight="normal"
