+++ client/src/components/Nnk.tsx
@@ -1,15 +1,14 @@
-import * as React from 'react';
+import React from 'react';
 import { useState, useEffect } from 'react';
 import { Loader } from './Loader';
 import { Card } from './Card';
 import '../App.css';
-import livedoorNews from './img/livedoorNews.png';
 
-export const Livedoor = () => {
-  const [data, setData] = useState([]);
+export const Nhk = () => {
+  const [data, setData] = useState<{ [key: string]: string }[]>([]);
 
   const getData = () => {
-    return fetch('./api/livedoor').then((response) => response.json());
+    return fetch('./api/nhk').then((response) => response.json());
   };
 
   useEffect(() => {
@@ -22,14 +21,14 @@ export const Livedoor = () => {
   return (
     <>
       <div className="inner">
-        <h2 className="section">Livedoor ニュース</h2>
+        <h2 className="section">NHK ニュース</h2>
         <div className="container">
           {data?.map((data) => (
             <Card
               key={data.textContent}
               title={data.textContent}
               link={data.href}
-              img={livedoorNews}
+              img={data.src}
               target="_blank"
               fontSize="1.5rem"
               fontWeight="normal"
