+++ client/src/components/Sankei.tsx
@@ -1,17 +1,17 @@
-import * as React from 'react';
+import React from 'react';
 import { useState, useEffect } from 'react';
 import { Loader } from './Loader';
 import { Card } from './Card';
 import '../App.css';
-import asahi from './img/asahi.png'; //modify here
+import sankei from './img/sankei.png'; //modify here
 
 //modify component Name
-export const Asahi = () => {
-  const [data, setData] = useState([]);
+export const Sankei = () => {
+  const [data, setData] = useState<{ [key: string]: string }[]>([]);
 
   //modify path
   const getData = () => {
-    return fetch('./api/asahi').then((response) => response.json());
+    return fetch('./api/sankei').then((response) => response.json());
   };
 
   useEffect(() => {
@@ -25,14 +25,14 @@ export const Asahi = () => {
   return (
     <>
       <div className="inner">
-        <h2 className="section">朝日新聞</h2>
+        <h2 className="section">産経新聞</h2>
         <div className="container">
           {data?.map((data) => (
             <Card
               key={data.textContent}
               title={data.textContent}
               link={data.href}
-              img={data.src}
+              img={data.src || sankei}
               target="_blank"
               fontSize="1.5rem"
               fontWeight="normal"
