+++ client/src/components/Yahoo.tsx
@@ -1,38 +1,35 @@
-import * as React from 'react';
+import React from 'react';
 import { useState, useEffect } from 'react';
 import { Loader } from './Loader';
 import { Card } from './Card';
 import '../App.css';
-import yomiuri from './img/yomiuri.png'; //modify here
+import yahooNews from './img/yahooNews.png';
 
-//modify component Name
-export const Yomiuri = () => {
-  const [data, setData] = useState([]);
+export const Yahoo = () => {
+  const [data, setData] = useState<{ [key: string]: string }[]>([]);
 
-  //modify path
   const getData = () => {
-    return fetch('./api/yomiuri').then((response) => response.json());
+    return fetch('./api/yahoo').then((response) => response.json());
   };
 
   useEffect(() => {
     getData().then((response) => setData(response));
   }, []);
 
-  //modify section text
   if (!data?.length) {
     return <Loader />;
   }
   return (
     <>
       <div className="inner">
-        <h2 className="section">読売新聞</h2>
+        <h2 className="section">Yahoo! ニュース</h2>
         <div className="container">
           {data?.map((data) => (
             <Card
               key={data.textContent}
               title={data.textContent}
               link={data.href}
-              img={yomiuri}
+              img={yahooNews}
               target="_blank"
               fontSize="1.5rem"
               fontWeight="normal"
