+++ client/src/components/Yomiuri.tsx
@@ -1,17 +1,17 @@
-import * as React from 'react';
+import React from 'react';
 import { useState, useEffect } from 'react';
 import { Loader } from './Loader';
 import { Card } from './Card';
 import '../App.css';
-import mainichi from './img/mainichi.png'; //modify here
+import yomiuri from './img/yomiuri.png'; //modify here
 
 //modify component Name
-export const Mainichi = () => {
-  const [data, setData] = useState([]);
+export const Yomiuri = () => {
+  const [data, setData] = useState<{ [key: string]: string }[]>([]);
 
   //modify path
   const getData = () => {
-    return fetch('./api/mainichi').then((response) => response.json());
+    return fetch('./api/yomiuri').then((response) => response.json());
   };
 
   useEffect(() => {
@@ -25,14 +25,14 @@ export const Mainichi = () => {
   return (
     <>
       <div className="inner">
-        <h2 className="section">毎日新聞</h2>
+        <h2 className="section">読売新聞</h2>
         <div className="container">
           {data?.map((data) => (
             <Card
               key={data.textContent}
               title={data.textContent}
               link={data.href}
-              img={mainichi}
+              img={yomiuri}
               target="_blank"
               fontSize="1.5rem"
               fontWeight="normal"
