+++ client/src/types.ts
@@ -1,0 +1,1 @@
+export type TODO_FROM_JS_TO_TS = any;
